+++ frontend/src/erd/__tests__/prisma.test.ts
@@ -63,8 +63,8 @@ describe('exportPrisma', () => {
         id: 'e1',
         source: '2',
         target: '1',
-        sourceHandle: 'src-user_id',
-        targetHandle: 'tgt-id',
+        sourceHandle: 'src-c-0075-0073-0065-0072-005f-0069-0064',
+        targetHandle: 'tgt-c-0069-0064',
         label: 'users_posts',
       },
     ];
@@ -224,8 +224,8 @@ describe('exportPrisma', () => {
         id: 'e1',
         source: '2',
         target: '1',
-        sourceHandle: 'src-user_id',
-        targetHandle: 'tgt-id',
+        sourceHandle: 'src-c-0075-0073-0065-0072-005f-0069-0064',
+        targetHandle: 'tgt-c-0069-0064',
         label: '1to1',
       },
     ];